+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Version 0.2.2
 
+- Added chart showing wins and losses by map type to the Trends page
+
 ## [Version 0.2.1](https://github.com/cheshire137/competiwatch-desktop/releases/tag/0.2.1)
 
 - Added chart to Trends tab showing wins and losses by whether or not you joined voice chat